PYTHON-5282 Move config utility functions to separate file
